tests: refactoring
